+++ DocumentRepoAPI/DocumentRepoAPI.Data/Entities/Folders.cs
@@ -21,9 +21,9 @@ namespace DocumentRepoAPI.Data.Entities
         public bool ReadAccess { get; set; }
         public bool ModifyAccess { get; set; }
         public bool DeleteAccess { get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public System.DateTime CreateDate { get; set; }
-        public int ModifiedBy { get; set; }
+        public long ModifiedBy { get; set; }
         public System.DateTime ModifiedDate { get; set; }
     }
 }
