+++ DocumentRepoAPI/DocumentRepoAPI.Data/Entities/SharedFiles.cs
@@ -18,9 +18,9 @@ namespace DocumentRepoAPI.Data.Entities
         public int SharedFile { get; set; }
         public int SharedBy { get; set; }
         public int SharedTo { get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public System.DateTime CreateDate { get; set; }
-        public int ModifiedBy { get; set; }
+        public long ModifiedBy { get; set; }
         public System.DateTime ModifiedDate { get; set; }
     }
 }
